advanced rearrange, finsihed 39, 45, 46
