+++ CHALLENGE_MAP.md
@@ -63,16 +63,16 @@
 
 - **Advanced Rendering**
 
-  - :pencil2: 38: Advanced JavaScript in `render` & `return` of React Component (Peter)
-  - :blue_book: 39: Render with an If/Else Condition
-  - :blue_book: 40: Use && for a More Concise Conditional
-  - :blue_book: 41: Return null to Prevent Rendering
-  - :pencil2: 42: Conditional rendering: ternary condition (Peter)
-  - :pencil2: 43: Conditional rendering based on `props`(Peter)
-  - :pencil2: 44: Use `map` to dynamically render an array of data (Peter)
-  - :blue_book: 45: Use Filter to Dynamically Filter an Array
-  - :blue_book: 46: Introducing Inline Styles
-  - :blue_book: 47: Add Inline Styles in React
+  - :blue_book: 38: Introducing Inline Styles
+  - :blue_book: 39: Add Inline Styles in React
+  - :pencil2: 40: Advanced JavaScript in `render` & `return` of React Component (Peter)
+  - :blue_book: 41: Render with an If/Else Condition
+  - :blue_book: 42: Use && for a More Concise Conditional
+  - :blue_book: 43: Return null to Prevent Rendering
+  - :pencil2: 44: Conditional rendering: ternary condition (Peter)
+  - :pencil2: 45: Conditional rendering based on `props`(Peter)
+  - :pencil2: 46: Use `map` to dynamically render an array of data (Peter)
+  - :blue_book: 47: Use Filter to Dynamically Filter an Array
   - :pencil2: 48: JSX Gotchas: Key difference in attribute names, etc (Peter)
   - :blue_book: 49: Render React on the Server with renderToString
