+++ src/challenges/react/curriculum/React_41.js
@@ -10,23 +10,22 @@ import { transform } from 'babel-standalone'
 export const QA = false;
 
 // ---------------------------- define challenge title ----------------------------
-export const challengeTitle = `<span class = 'default'>Challenge: </span>Return null to Prevent Rendering`
+export const challengeTitle = `<span class = 'default'>Challenge: </span>Render with an If/Else Condition`
 
 // ---------------------------- challenge text ----------------------------
-export const challengeText = `<span class = 'default'>Intro: </span>In some cases we may not want to render a child element under certain conditions. In React you can avoid
-rendering a component by simply returning <code>null</code> in its <code>render()</code> method. This can be used in some cases for example where you have a child
-component that just doesn't need to render under some conditions.`
+export const challengeText = `<span class = 'default'>Intro: </span>One application of using JavaScript to control our rendered view is to create a simple condition
+and only render some elements when this condition is true. Here we will do this with a standard <code>if/else</code> statement in the <code>render()</code> method
+of a React component.`
 
 // ---------------------------- challenge instructions ----------------------------
-export const challengeInstructions = `<span class = 'default'>Instructions: </span>We've provided the example we've been working with again here. This time, we will render
-our <code>h1</code> element in the <code>Child</code> component rather than in one top-level component. We will pass our <code>display</code> condition as <code>props</code>
-to this child component. We can check the condition of <code>display</code> in the child component and if it is <code>false</code> we can return <code>null</code>. Try it out!
-Now we've accomplished the same behavior in three different ways. This should begin to show you the versatility of using JavaScript to write our UIs. React provides us a lot
-of control over what and how we render our views.`
+export const challengeInstructions = `<span class = 'default'>Instructions: </span>We've created a component that contains a <code>boolean</code> in its state
+which represents a condition which tracks if we want to display some element in the UI or not. We've wired up a <code>&lt;button/&gt;</code> to toggle the state of this value. Currently,
+we render the same UI everytime. Rewrite the <code>render()</code> method with an <code>if/else</code> statement so that if <code>display</code> is <code>true</code> we return the current
+markup, otherwise, let's just return the <code>&lt;button/&gt;</code> without the <code>&lt;h1/&gt;</code> element.`
 
 // ---------------------------- define challenge seed code ----------------------------
 export const seedCode =
-`class Parent extends React.Component {
+`class MyComponent extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
@@ -43,23 +42,15 @@ export const seedCode =
     return (
 	   	<div>
 	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
-	   		<Child display={this.state.display}/>
+	   		<h1>Displayed!</h1>
 	   	</div>
     );
   }
-};
-
-class Child extends React.Component {
-	constructor(props) {
-		super(props);
-	}
-	render() {
-		// change code below this line
-}`
+};`
 
 // ---------------------------- define challenge solution code ----------------------------
 export const solutionCode =
-`class Parent extends React.Component {
+`class MyComponent extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
@@ -73,39 +64,32 @@ export const solutionCode =
 	}
   render() {
   	// change code below this line
-    return (
-	   	<div>
-	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
-	   		<Child display={this.state.display}/>
-	   	</div>
-    );
+  	if (this.state.display) {
+	    return (
+		   	<div>
+		   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+		   		<h1>Displayed!</h1>
+		   	</div>
+	    );
+	  } else {
+	  	return (
+	  		<div>
+		   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+		   	</div>
+	  	);
+	  }
   }
-};
-
-class Child extends React.Component {
-	constructor(props) {
-		super(props);
-	}
-	render() {
-		// change code below this line
-		if (this.props.display) {
-			return <h1>Display!</h1>
-		} else {
-			return null;
-		}
-	}
-}`
+};`
 
 // ---------------------------- define challenge tests ----------------------------
 
 export const executeTests = (code) => {
 
 	const error_0 = 'Your JSX code was transpiled successfully.';
-	const error_1 = 'The Parent component exists and is rendered.';
-	const error_2 = 'The Child component exists and is rendered.';
-	const error_3 = 'When display is set to true, a div, button, and h1 are rendered.';
-	const error_4 = 'When display is set to false, only a div and button are rendered.';
-	const error_5 = 'The Child component returns null when passed a falsy value for the display prop, otherwise it returns an h1 element.';
+	const error_1 = 'MyComponent exists and is rendered.';
+	const error_2 = 'When display is set to true, a div, button, and h1 are rendered.';
+	const error_3 = 'When display is set to false, only a div and button are rendered.';
+	const error_4 = 'The render method uses and if/else statement to check the condition of this.state.display.';
 
 	let testResults = [
 		{
@@ -132,11 +116,6 @@ export const executeTests = (code) => {
 			test: 4,
 			status: false,
 			condition: error_4
-		},
-		{
-			test: 5,
-			status: false,
-			condition: error_5
 		}
 	];
 
@@ -144,7 +123,7 @@ export const executeTests = (code) => {
 
 	// this applies an export to the user's code so
 	// we can access their component here for tests
-	const exportScript = '\n export default Parent'
+	const exportScript = '\n export default MyComponent'
 	const modifiedCode = code.concat(exportScript);
 	
 	// test 0: try to transpile JSX, ES6 code to ES5 in browser
@@ -170,7 +149,7 @@ export const executeTests = (code) => {
 
 	// test 1:
 	try {
-		assert.strictEqual(mockedComponent.find('Parent').length, 1, error_1);
+		assert.strictEqual(mockedComponent.find('MyComponent').length, 1, error_1);
 		testResults[1].status = true;
 	} catch (err) {
 		passed = false;
@@ -179,61 +158,47 @@ export const executeTests = (code) => {
 
 	// test 2:
 	try {
-		assert.strictEqual(mockedComponent.find('Child').length, 1, error_2);
-		testResults[2].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[2].status = false;
-	}	
-
-	// test 3:
-	try {
 		mockedComponent.setState({display: true});
 		assert(
 			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('div').children().length === 2 &&
 			mockedComponent.find('button').length === 1 &&
 			mockedComponent.find('h1').length === 1,
+			error_2
+		);
+		testResults[2].status = true;
+	} catch (err) {
+		passed = false;
+		testResults[2].status = false;		
+	}
+
+	// test 3:
+	try {
+		mockedComponent.setState({display: false});
+		assert(
+			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('div').children().length === 1 &&
+			mockedComponent.find('button').length === 1 &&
+			mockedComponent.find('h1').length === 0,
 			error_3
 		);
 		testResults[3].status = true;
 	} catch (err) {
 		passed = false;
-		testResults[3].status = false;		
+		testResults[3].status = false;
 	}
-
-	let displayFalse, displayTrue;
 
 	// test 4:
 	try {
-		mockedComponent.setState({display: false});
 		assert(
-			mockedComponent.find('div').length === 1 &&
-			mockedComponent.find('button').length === 1 &&
-			mockedComponent.find('h1').length === 0,
+			code.includes('if') === true &&
+			code.includes('else') === true,
 			error_4
 		);
 		testResults[4].status = true;
 	} catch (err) {
 		passed = false;
 		testResults[4].status = false;
-	}
-
-	// test 5:
-	try {
-		mockedComponent.setState({display: true});
-		displayTrue = mockedComponent.find('Child').find('h1');
-		mockedComponent.setState({display: false});
-		displayFalse = mockedComponent.find('Child').find('h1');
-		assert(
-			code.includes('null') === true &&
-			displayTrue.length === 1 &&
-			displayFalse.length === 0,
-			error_5
-		);
-		testResults[5].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[5].status = false;
 	}
 
 	return {
@@ -248,7 +213,7 @@ export const executeTests = (code) => {
 export const liveRender = (code) => {
 
 	try {
-		const exportScript = '\n export default Parent'
+		const exportScript = '\n export default MyComponent'
 		const modifiedCode = code.concat(exportScript);
 		const es5 = transform(modifiedCode, { presets: [ 'es2015', 'stage-2', 'react' ] }).code;
 		const renderedComponent = React.createElement(eval(es5));
