+++ src/challenges/react/curriculum/React_42.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import React from 'react'
 import assert from 'assert'
-import { shallow, mount } from 'enzyme'
+import { mount } from 'enzyme'
 import { transform } from 'babel-standalone'
 
 // snippet for defining HTML: <code>&lt;div /&gt;</code>
@@ -10,103 +10,72 @@ import { transform } from 'babel-standalone'
 export const QA = false;
 
 // ---------------------------- define challenge title ----------------------------
-export const challengeTitle = `<span class = 'default'>Challenge: </span>Conditional Rendering with the Ternary Operator`
+export const challengeTitle = `<span class = 'default'>Challenge: </span>Use && for a More Concise Conditional`
 
 // ---------------------------- challenge text ----------------------------
-export const challengeText = `<span class = 'default'>Intro: </span>_CHALLENGE_TEXT_`
+export const challengeText = `<span class = 'default'>Intro: </span>Nice job! Now we will demonstrate a more concise way to achieve the same result. Imagine that we are
+tracking several conditions in our component and we want to respond differently to each of these conditions. We don't want to keep writing <code>else if</code> conditions to
+just return slightly tweaked UIs. Instead, we can use the <code>&&</code> logical operator to perform conditional logic in a more concise way. This
+is possible because we want to check if a condition is <code>true</code>, and if it is return some markup. So if we write:<br><br>
+
+<code>{condition && &lt;p&gt;markup&lt;/p&gt;}</code><br><br>
+
+If the <code>condition</code> is <code>true</code> the markup will be returned, otherwise if it is <code>false</code> the operation will immediately
+return <code>false</code> after evaluating the <code>condition</code> and return nothing. We can include these statments directly in our JSX
+and even string multiple conditions together by continuing to write <code>&&</code> after each check we make. This allows us to 
+handle more complex conditional logic in our <code>render()</code> method.`
 
 // ---------------------------- challenge instructions ----------------------------
-export const challengeInstructions = `<span class = 'default'>Instructions: </span>_ADD_YOUR_INSTRUCTIONS_HERE_`
+export const challengeInstructions = `<span class = 'default'>Instructions: </span>Solve the previous example again, this time using the <code>&&</code> logical operator.`
 
 // ---------------------------- define challenge seed code ----------------------------
-export const seedCode = `
-const inputStyle = {
-	width: 235,
-	margin: 5
-}
-
-class CheckUserAge extends React.Component {
-	constructor(props) {
-		super(props);
-		// change code below this line 
-
-		// change code above this line
-		this.submit = this.submit.bind(this);
-		this.handleChange = this.handleChange.bind(this);
-	}
-	handleChange(e) {
-		this.setState({
-			input: e.target.value,
-			userAge: ''
-		});
-	}
-	submit() {
-		this.setState({
-			userAge: this.state.input
-		});
-	}
-	render() {
-		const buttonText = ["Submit", "You May Enter", "You Shall Not Pass"];
-		const button = // change code here
-		return (
-			<div>
-				<h3>Enter Your Age to Continue</h3>
-				<input 
-					style={inputStyle}
-					type="number"
-					value={this.state.input}
-					onChange={this.handleChange} /><br />
-				{button}
-			</div>
-		);
-	}
-};`
-
-// ---------------------------- define challenge solution code ----------------------------
-export const solutionCode = `
-const inputStyle = {
-	width: 235,
-	margin: 5
-}
-
-class CheckUserAge extends React.Component {
+export const seedCode =
+`class MyComponent extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
-			userAge: '',
-			input: '',
+			display: true
 		}
-		this.submit = this.submit.bind(this);
-		this.handleChange = this.handleChange.bind(this);
 	}
-	handleChange(e) {
+	toggleDisplay = () => {
 		this.setState({
-			input: e.target.value,
-			userAge: ''
+			display: !this.state.display
 		});
 	}
-	submit() {
+  render() {
+  	// change code below this line
+    return (
+	   	<div>
+	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+	   		<h1>Displayed!</h1>
+	   	</div>
+    );
+  }
+};`
+
+// ---------------------------- define challenge solution code ----------------------------
+export const solutionCode =
+`class MyComponent extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
+			display: true
+		}
+	}
+	toggleDisplay = () => {
 		this.setState({
-			userAge: this.state.input
+			display: !this.state.display
 		});
 	}
-	render() {
-		const buttonText = ["Submit", "You May Enter", "You Shall Not Pass"];
-		const button = this.state.userAge === '' ? <button onClick={this.submit}>{buttonText[0]}</button> :
-			this.state.userAge >= 18 ? <button>{buttonText[1]}</button> : 
-			<button>{buttonText[2]}</button>;
-		return (
-			<div>
-				<h3>Enter Your Age to Continue</h3>
-				<input 
-					style={inputStyle}
-					type="number"
-					value={this.state.input}
-					onChange={this.handleChange} /><br />
-				{button}
-			</div>
-		);
-	}
+  render() {
+  	// change code below this line
+    return (
+	   	<div>
+	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+	   		{this.state.display && <h1>Displayed!</h1>}
+	   	</div>
+    );
+  }
 };`
 
 // ---------------------------- define challenge tests ----------------------------
@@ -114,15 +83,10 @@ class CheckUserAge extends React.Component {
 export const executeTests = (code) => {
 
 	const error_0 = 'Your JSX code was transpiled successfully.';
-	const error_1 = 'The CheckUserAge component exists and is rendered to the page.';
-	const error_2 = 'The CheckUserAge component returns a single <input> element.';
-	const error_3 = 'The CheckUserAge component returns a single <button> element.';
-	const error_4 = 'The CheckUserAge component\'s state is initialized with a property of userAge and a property of input, both set to a value of an empty string.';
-	const error_5 = 'When the CheckUserAge component is first rendered to the DOM, the button\'s inner text should read "Submit".';
-	const error_6 = 'When a number of less than 18 is entered into the <input> element and the button is clicked, the button\'s inner text should read "You Shall Not Pass".';
-	const error_7 = 'When a number greater than or equal to 18 is entered into the <input> element and the button is clicked, the button\'s inner text should read "You May Enter".';
-	const error_8 = 'Once a number has been submited, and the value of the input is once again changed, the button should return to reading "Submit".'
-
+	const error_1 = 'MyComponent exists and is rendered.';
+	const error_2 = 'When display is set to true, a div, button, and h1 are rendered.';
+	const error_3 = 'When display is set to false, only a div and button are rendered.';
+	const error_4 = 'The render method uses the && logical operator to check the condition of this.state.display.';
 
 	let testResults = [
 		{
@@ -146,29 +110,9 @@ export const executeTests = (code) => {
 			condition: error_3
 		},
 		{
-			test: 3,
+			test: 4,
 			status: false,
 			condition: error_4
-		},
-		{
-			test: 3,
-			status: false,
-			condition: error_5
-		},
-		{
-			test: 3,
-			status: false,
-			condition: error_6
-		},
-		{
-			test: 3,
-			status: false,
-			condition: error_7
-		},
-		{
-			test: 3,
-			status: false,
-			condition: error_8
 		}
 	];
 
@@ -176,12 +120,12 @@ export const executeTests = (code) => {
 
 	// this applies an export to the user's code so
 	// we can access their component here for tests
-	const exportScript = '\n export default CheckUserAge'
+	const exportScript = '\n export default MyComponent'
 	const modifiedCode = code.concat(exportScript);
 	
 	// test 0: try to transpile JSX, ES6 code to ES5 in browser
 	try {
-		es5 = transform(modifiedCode, { presets: [ 'es2015', 'react' ] }).code;
+		es5 = transform(modifiedCode, { presets: [ 'es2015', 'stage-2', 'react' ] }).code;
 		testResults[0].status = true;
 	} catch (err) {
 		passed = false;
@@ -197,16 +141,12 @@ export const executeTests = (code) => {
 		passed = false;
 	}
 
-	console.log(mockedComponent);
-
-	let initialState, state_1, state_2, state_3, state_4;
-
 	// run specific tests to verify the functionality
 	// that the challenge is trying to assess:
 
 	// test 1:
 	try {
-		assert.strictEqual(mockedComponent.find('CheckUserAge').length, 1, error_1);
+		assert.strictEqual(mockedComponent.find('MyComponent').length, 1, error_1);
 		testResults[1].status = true;
 	} catch (err) {
 		passed = false;
@@ -215,7 +155,14 @@ export const executeTests = (code) => {
 
 	// test 2:
 	try {
-		assert.strictEqual(mockedComponent.find('input').length, 1, error_2);
+		mockedComponent.setState({display: true});
+		assert(
+			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('div').children().length === 2 &&
+			mockedComponent.find('button').length === 1 &&
+			mockedComponent.find('h1').length === 1,
+			error_2
+		);
 		testResults[2].status = true;
 	} catch (err) {
 		passed = false;
@@ -224,7 +171,14 @@ export const executeTests = (code) => {
 
 	// test 3:
 	try {
-		assert.strictEqual(mockedComponent.find('button').length, 1, error_3);
+		mockedComponent.setState({display: false});
+		assert(
+			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('div').children().length === 1 &&
+			mockedComponent.find('button').length === 1 &&
+			mockedComponent.find('h1').length === 0,
+			error_3
+		);
 		testResults[3].status = true;
 	} catch (err) {
 		passed = false;
@@ -233,92 +187,11 @@ export const executeTests = (code) => {
 
 	// test 4:
 	try {
-		initialState = mockedComponent.state();
-		assert(initialState.userAge === '' && initialState.input === '', error_4);
+		assert.strictEqual(code.includes('&&'), true, error_4);
 		testResults[4].status = true;
 	} catch (err) {
 		passed = false;
 		testResults[4].status = false;
-	}
-
-	// test 5:
-	try {
-		assert.strictEqual(mockedComponent.find('button').nodes[0].innerText, "Submit", error_5);
-		testResults[5].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[5].status = false;
-	}
-
-	// test 6:
-	try {
-		mockedComponent.find('input').simulate('change', {target: { value: 0 }});
-		mockedComponent.find('button').simulate('click');
-		state_1 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 10 }});
-		mockedComponent.find('button').simulate('click');
-		state_2 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 17 }});
-		mockedComponent.find('button').simulate('click');
-		state_3 = mockedComponent.find('button').nodes[0].innerText;
-		assert(
-			state_1 === 'You Shall Not Pass' &&
-			state_2 === 'You Shall Not Pass' &&
-			state_3 === 'You Shall Not Pass',
-			error_6
-		);
-		testResults[6].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[6].status = false;
-	}
-
-	// test 7:
-	try {
-		mockedComponent.find('input').simulate('change', {target: { value: 18 }});
-		mockedComponent.find('button').simulate('click');
-		state_1 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 25 }});
-		mockedComponent.find('button').simulate('click');
-		state_2 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 35 }});
-		mockedComponent.find('button').simulate('click');
-		state_3 = mockedComponent.find('button').nodes[0].innerText;
-		assert(
-			state_1 === 'You May Enter' &&
-			state_2 === 'You May Enter' &&
-			state_3 === 'You May Enter',
-			error_7
-		);
-		testResults[7].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[7].status = false;
-	}
-
-	// test 8:
-	try {
-		mockedComponent.find('input').simulate('change', {target: { value: 18 }});
-		mockedComponent.find('button').simulate('click');
-		state_1 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 1 }});
-		state_2 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 10 }});
-		mockedComponent.find('button').simulate('click');
-		state_3 = mockedComponent.find('button').nodes[0].innerText;
-		mockedComponent.find('input').simulate('change', {target: { value: 1 }});
-		state_4 = mockedComponent.find('button').nodes[0].innerText;
-		assert(
-			state_1 === 'You May Enter' &&
-			state_2 === 'Submit' &&
-			state_3 === 'You Shall Not Pass' &&
-			state_4 === 'Submit',
-			error_8
-		);
-		testResults[8].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[8].status = false;
 	}
 
 	return {
@@ -333,9 +206,9 @@ export const executeTests = (code) => {
 export const liveRender = (code) => {
 
 	try {
-		const exportScript = '\n export default CheckUserAge'
+		const exportScript = '\n export default MyComponent'
 		const modifiedCode = code.concat(exportScript);
-		const es5 = transform(modifiedCode, { presets: [ 'es2015', 'react' ] }).code;
+		const es5 = transform(modifiedCode, { presets: [ 'es2015', 'stage-2', 'react' ] }).code;
 		const renderedComponent = React.createElement(eval(es5));
 		return renderedComponent;
 	} catch (err) {
