+++ src/challenges/react/curriculum/React_43.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import React from 'react'
 import assert from 'assert'
-import { shallow, mount } from 'enzyme'
+import { mount } from 'enzyme'
 import { transform } from 'babel-standalone'
 
 // snippet for defining HTML: <code>&lt;div /&gt;</code>
@@ -10,113 +10,102 @@ import { transform } from 'babel-standalone'
 export const QA = false;
 
 // ---------------------------- define challenge title ----------------------------
-export const challengeTitle = `<span class = 'default'>Challenge: </span>Conditional Rendering with Props`
+export const challengeTitle = `<span class = 'default'>Challenge: </span>Return null to Prevent Rendering`
 
 // ---------------------------- challenge text ----------------------------
-export const challengeText = `<span class = 'default'>Intro: </span>_CHALLENGE_TEXT_`
+export const challengeText = `<span class = 'default'>Intro: </span>In some cases we may not want to render a child element under certain conditions. In React you can avoid
+rendering a component by simply returning <code>null</code> in its <code>render()</code> method. This can be used in some cases for example where you have a child
+component that just doesn't need to render under some conditions.`
 
 // ---------------------------- challenge instructions ----------------------------
-export const challengeInstructions = `<span class = 'default'>Instructions: </span>_ADD_YOUR_INSTRUCTIONS_HERE_`
+export const challengeInstructions = `<span class = 'default'>Instructions: </span>We've provided the example we've been working with again here. This time, we will render
+our <code>h1</code> element in the <code>Child</code> component rather than in one top-level component. We will pass our <code>display</code> condition as <code>props</code>
+to this child component. We can check the condition of <code>display</code> in the child component and if it is <code>false</code> we can return <code>null</code>. Try it out!
+Now we've accomplished the same behavior in three different ways. This should begin to show you the versatility of using JavaScript to write our UIs. React provides us a lot
+of control over what and how we render our views.`
 
 // ---------------------------- define challenge seed code ----------------------------
-export const seedCode = `
-class Result extends React.Component {
-	constructor(props) {
-		super(props);
-	}
-	render() {
-		return (
-			<h1>
-			{
-				/* change code here */
-			}
-			</h1>
-		)
-	};
-};
-
-class GameOfChance extends React.Component {
+export const seedCode =
+`class Parent extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
-			counter: 1
+			display: true
 		}
-		this.handleClick = this.handleClick.bind(this);
 	}
-	handleClick() {
+	toggleDisplay = () => {
 		this.setState({
-			// change code here
+			display: !this.state.display
 		});
 	}
-	render() {
-		return (
-			<div>
-				<button onClick={this.handleClick}>Play Again</button>
-				{ /* change code below this line /* }
+  render() {
+  	// change code below this line
+    return (
+	   	<div>
+	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+	   		<Child display={this.state.display}/>
+	   	</div>
+    );
+  }
+};
 
-				{ /* change code above this line /* }
-				<p>{'Turn: ' + this.state.counter}</p>
-			</div>
-		);
+class Child extends React.Component {
+	constructor(props) {
+		super(props);
 	}
-};`
+	render() {
+		// change code below this line
+}`
 
 // ---------------------------- define challenge solution code ----------------------------
-export const solutionCode = `
-class Result extends React.Component {
-	constructor(props) {
-		super(props);
-	}
-	render() {
-		return (
-			<h1>
-			{
-				this.props.fiftyFifty > .5 ? 
-				'You Win!' :
-				'You Lose!'
-			}
-			</h1>
-		)
-	};
-};
-
-class GameOfChance extends React.Component {
+export const solutionCode =
+`class Parent extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
-			counter: 1
+			display: true
 		}
-		this.handleClick = this.handleClick.bind(this);
 	}
-	handleClick() {
+	toggleDisplay = () => {
 		this.setState({
-			counter: this.state.counter + 1
+			display: !this.state.display
 		});
 	}
-	render() {
-		return (
-			<div>
-				<button onClick={this.handleClick}>Play Again</button>
-				<Result fiftyFifty={Math.random()} />
-				<p>{'Turn: ' + this.state.counter}</p>
-			</div>
-		);
+  render() {
+  	// change code below this line
+    return (
+	   	<div>
+	   		<button onClick={this.toggleDisplay}>Toggle Display</button>
+	   		<Child display={this.state.display}/>
+	   	</div>
+    );
+  }
+};
+
+class Child extends React.Component {
+	constructor(props) {
+		super(props);
 	}
-};`
+	render() {
+		// change code below this line
+		if (this.props.display) {
+			return <h1>Display!</h1>
+		} else {
+			return null;
+		}
+	}
+}`
 
 // ---------------------------- define challenge tests ----------------------------
 
 export const executeTests = (code) => {
 
 	const error_0 = 'Your JSX code was transpiled successfully.';
-	const error_1 = 'The GameOfChance component exists and is rendered to the page.';
-	const error_2 = 'GameOfChance should return a single <button> element.';
-	const error_3 = 'GameOfChance should return a single instance of the Result component.';
-	const error_4 = 'GameOfChance\'s state is initialized with a property of counter set to a value of 1.'
-	const error_5 = 'When the GameOfChance component is first rendered to the DOM, a <p> element should be returned with the inner text of "Turn: 1".'
-	const error_6 = 'Each time the buton is clicked, the counter state should be incremented by a value of 1, and a single <p> element should be rendered to the DOM that contains the text "Turn: N", where N is the value of the counter state.'
-	const error_7 = 'When the GameOfChance component is first mounted to the DOM and each time the button is clicked thereafter, a single <h1> element should be returned that randomly renders either "You Win!" or "You Lose!".';
-
+	const error_1 = 'The Parent component exists and is rendered.';
+	const error_2 = 'The Child component exists and is rendered.';
+	const error_3 = 'When display is set to true, a div, button, and h1 are rendered.';
+	const error_4 = 'When display is set to false, only a div and button are rendered.';
+	const error_5 = 'The Child component returns null when passed a falsy value for the display prop, otherwise it returns an h1 element.';
 
 	let testResults = [
 		{
@@ -148,16 +137,6 @@ export const executeTests = (code) => {
 			test: 5,
 			status: false,
 			condition: error_5
-		},
-		{
-			test: 6,
-			status: false,
-			condition: error_6
-		},
-		{
-			test: 7,
-			status: false,
-			condition: error_7
 		}
 	];
 
@@ -165,15 +144,12 @@ export const executeTests = (code) => {
 
 	// this applies an export to the user's code so
 	// we can access their component here for tests
-	const exportScript = '\n export default GameOfChance'
+	const exportScript = '\n export default Parent'
 	const modifiedCode = code.concat(exportScript);
-
-	let initialState, counter_1, counter_2, counter_3, counter_4, counter_5, counter_6, counter_7, counter_8, counter_9, counter_10;
-	let p_1, p_2, p_3, p_4, p_5, p_6, p_7, p_8, p_9, p_10;
 	
 	// test 0: try to transpile JSX, ES6 code to ES5 in browser
 	try {
-		es5 = transform(modifiedCode, { presets: [ 'es2015', 'react' ] }).code;
+		es5 = transform(modifiedCode, { presets: [ 'es2015', 'stage-2', 'react' ] }).code;
 		testResults[0].status = true;
 	} catch (err) {
 		passed = false;
@@ -189,14 +165,12 @@ export const executeTests = (code) => {
 		passed = false;
 	}
 
-	console.log(mockedComponent.instance());
-
 	// run specific tests to verify the functionality
 	// that the challenge is trying to assess:
 
 	// test 1:
 	try {
-		assert.strictEqual(mockedComponent.find('GameOfChance').length, 1, error_1)
+		assert.strictEqual(mockedComponent.find('Parent').length, 1, error_1);
 		testResults[1].status = true;
 	} catch (err) {
 		passed = false;
@@ -205,75 +179,61 @@ export const executeTests = (code) => {
 
 	// test 2:
 	try {
-		assert.strictEqual(mockedComponent.find('button').length, 1, error_2)
+		assert.strictEqual(mockedComponent.find('Child').length, 1, error_2);
 		testResults[2].status = true;
 	} catch (err) {
 		passed = false;
-		testResults[2].status = false;		
-	}
+		testResults[2].status = false;
+	}	
 
 	// test 3:
 	try {
-		assert.strictEqual(mockedComponent.find('Result').length, 1, error_3)
+		mockedComponent.setState({display: true});
+		assert(
+			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('button').length === 1 &&
+			mockedComponent.find('h1').length === 1,
+			error_3
+		);
 		testResults[3].status = true;
 	} catch (err) {
 		passed = false;
-		testResults[3].status = false;
+		testResults[3].status = false;		
 	}
 
+	let displayFalse, displayTrue;
+
+	// test 4:
 	try {
-		initialState = mockedComponent.state();
-		assert.strictEqual(initialState.counter, 1, error_4);
+		mockedComponent.setState({display: false});
+		assert(
+			mockedComponent.find('div').length === 1 &&
+			mockedComponent.find('button').length === 1 &&
+			mockedComponent.find('h1').length === 0,
+			error_4
+		);
 		testResults[4].status = true;
 	} catch (err) {
 		passed = false;
 		testResults[4].status = false;
 	}
 
+	// test 5:
 	try {
-		assert.strictEqual(mockedComponent.find('p').nodes[0].innerText, "Turn: 1", error_5);
+		mockedComponent.setState({display: true});
+		displayTrue = mockedComponent.find('Child').find('h1');
+		mockedComponent.setState({display: false});
+		displayFalse = mockedComponent.find('Child').find('h1');
+		assert(
+			code.includes('null') === true &&
+			displayTrue.length === 1 &&
+			displayFalse.length === 0,
+			error_5
+		);
 		testResults[5].status = true;
 	} catch (err) {
 		passed = false;
 		testResults[5].status = false;
-	}
-
-	try {
-		mockedComponent.find('button').simulate('click');
-		counter_1 = mockedComponent.state().counter;
-		p_1 = mockedComponent.find('p').nodes[0].innerText
-		mockedComponent.find('button').simulate('click');
-		counter_2 = mockedComponent.state().counter;
-		p_2 = mockedComponent.find('p').nodes[0].innerText
-		mockedComponent.find('button').simulate('click');
-		counter_3 = mockedComponent.state().counter;
-		p_3 = mockedComponent.find('p').nodes[0].innerText
-		mockedComponent.find('button').simulate('click');
-		counter_4 = mockedComponent.state().counter;
-		p_4 = mockedComponent.find('p').nodes[0].innerText
-		mockedComponent.find('button').simulate('click');
-		counter_5 = mockedComponent.state().counter;
-		p_5 = mockedComponent.find('p').nodes[0].innerText
-		assert(
-			counter_1 === 2 && p_1 === 'Turn: 2' &&
-			counter_2 === 3 && p_2 === 'Turn: 3' &&
-			counter_3 === 4 && p_3 === 'Turn: 4' &&
-			counter_4 === 5 && p_4 === 'Turn: 5' &&
-			counter_5 === 6 && p_5 === 'Turn: 6',
-			error_6
-		);
-		testResults[6].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[6].status = false;
-	}
-
-	try {
-		assert();
-		testResults[7].status = true;
-	} catch (err) {
-		passed = false;
-		testResults[7].status = false;
 	}
 
 	return {
@@ -288,9 +248,9 @@ export const executeTests = (code) => {
 export const liveRender = (code) => {
 
 	try {
-		const exportScript = '\n export default GameOfChance'
+		const exportScript = '\n export default Parent'
 		const modifiedCode = code.concat(exportScript);
-		const es5 = transform(modifiedCode, { presets: [ 'es2015', 'react' ] }).code;
+		const es5 = transform(modifiedCode, { presets: [ 'es2015', 'stage-2', 'react' ] }).code;
 		const renderedComponent = React.createElement(eval(es5));
 		return renderedComponent;
 	} catch (err) {
